`beautify`, `beautify_to_string` helper functions
